Backend/Frontend: Set up axios, call the movies api using useEffect, and avoid CORS errors when making requests (01:57:25)
Prevent browsers from showing CORS errors when making preflight requests by adding "Access-Control-Allow-Origin: *" in the response header. In Spring Boot, we just have to annotate controllers with @CrossOrigin to add that.

Note: The default frontend debug server is bound to localhost:3000, and the backend is localhost:8080. Although they have the same host name, they are still not considered the same origin becase the ports are different. That's why we may see CORS errors if we do not add the access control in the resposne header sent by the backend.