Use CRA proxy to aoivd the CORS problem
Create react app provides a way to set up a proxy to the backend. By doing that, requests will be sent to the same origin (same host and same port) for both the react page and the backend API. So, we do not have to enable the cross-origin access for all routes in the backend.